[Missions] add filters on map
