+++ frontend/src/components/style.ts
@@ -10,10 +10,11 @@ export const Bold = styled.span`
 `
 
 export const StyledMapMenuDialogContainer = styled(MapMenuDialog.Container)`
+  display: block;
+  margin-left: -6px;
+  max-height: auto;
   position: absolute;
   transform: translateX(-100%);
-  margin-left: -6px;
-  display: block;
 `
 
 // TODO delete when Monitor-ui component have good padding
