+++ frontend/src/features/missions/components/Filters/Table/FilterTags.tsx
@@ -1,19 +1,14 @@
+import { RTK_DEFAULT_QUERY_OPTIONS } from '@api/constants'
+import { useGetControlUnitsQuery } from '@api/controlUnitsAPI'
+import { useAppDispatch } from '@hooks/useAppDispatch'
+import { useAppSelector } from '@hooks/useAppSelector'
+import { useGetControlPlans } from '@hooks/useGetControlPlans'
+import { FrontendError } from '@libs/FrontendError'
 import { SingleTag } from '@mtes-mct/monitor-ui'
+import { FrontCompletionStatusLabel, missionStatusLabels, missionTypeEnum } from 'domain/entities/missions'
+import { MissionFiltersEnum, updateFilters, type MissionFiltersState } from 'domain/shared_slices/MissionFilters'
 import { useMemo } from 'react'
 import styled from 'styled-components'
-
-import { RTK_DEFAULT_QUERY_OPTIONS } from '../../../../api/constants'
-import { useGetControlUnitsQuery } from '../../../../api/controlUnitsAPI'
-import { FrontCompletionStatusLabel, missionStatusLabels, missionTypeEnum } from '../../../../domain/entities/missions'
-import {
-  MissionFiltersEnum,
-  updateFilters,
-  type MissionFiltersState
-} from '../../../../domain/shared_slices/MissionFilters'
-import { useAppDispatch } from '../../../../hooks/useAppDispatch'
-import { useAppSelector } from '../../../../hooks/useAppSelector'
-import { useGetControlPlans } from '../../../../hooks/useGetControlPlans'
-import { FrontendError } from '../../../../libs/FrontendError'
 
 export function FilterTags() {
   const dispatch = useAppDispatch()
